+++ risk/model_enum_flow_type.go
@@ -15,17 +15,25 @@ import (
 	"fmt"
 )
 
-// EnumFlowType A string that specifies the flow type for the event. The only option (and default) is `AUTHENTICATION`.
+// EnumFlowType  The type of flow for which the risk evaluation is being carried out. Can take any of the following values:   `REGISTRATION` - initial registration of an account   `AUTHENTICATION` - standard authentication for login or actions such as password change   `ACCESS` - verification of whether user can access the relevant application, using tools such as PingAccess   `AUTHORIZATION` - verification of whether user is authorized to perform a specific action such as a profile change, using tools such as PingOne Authorize   `TRANSACTION` - authentication carried out in the context of a purchase or other monetary transaction   The default value is `AUTHENTICATION`. 
 type EnumFlowType string
 
 // List of EnumFlowType
 const (
 	ENUMFLOWTYPE_AUTHENTICATION EnumFlowType = "AUTHENTICATION"
+	ENUMFLOWTYPE_REGISTRATION EnumFlowType = "REGISTRATION"
+	ENUMFLOWTYPE_ACCESS EnumFlowType = "ACCESS"
+	ENUMFLOWTYPE_AUTHORIZATION EnumFlowType = "AUTHORIZATION"
+	ENUMFLOWTYPE_TRANSACTION EnumFlowType = "TRANSACTION"
 )
 
 // All allowed values of EnumFlowType enum
 var AllowedEnumFlowTypeEnumValues = []EnumFlowType{
 	"AUTHENTICATION",
+	"REGISTRATION",
+	"ACCESS",
+	"AUTHORIZATION",
+	"TRANSACTION",
 }
 
 func (v *EnumFlowType) UnmarshalJSON(src []byte) error {